woops needed to separate out this bash block

+++ cookbooks/rightimage/recipes/cloud_add_euca.rb
@@ -90,7 +90,6 @@ bash "install euca tools for centos" do
 end
 
 bash "configure for eucalyptus" do
-  only_if { node[:platform] == "centos" }
   flags "-ex"
   code <<-EOH
     guest_root=#{guest_root}
@@ -98,13 +97,19 @@ bash "configure for eucalyptus" do
     ## insert cloud file
     mkdir -p $guest_root/etc/rightscale.d
     echo -n "eucalyptus" > $guest_root/etc/rightscale.d/cloud
+  EOH
+end
 
+bash "clean out yum cache" do
+  only_if { node[:platform] == "centos" }
+  flags "-ex"
+  code <<-EOH
+    guest_root=#{guest_root}
     # clean out packages
     chroot $guest_root yum -y clean all
     
     rm ${guest_root}/var/lib/rpm/__*
     chroot $guest_root rpm --rebuilddb
-
   EOH
 end
